extracted players
